+++ src/deprecations.jl
@@ -1,6 +1,8 @@
 # Some deprecations exist in submodules Pathfinding, OSM
 
 @deprecate edistance euclidean_distance
+@deprecate rem_node! rem_vertex! 
+@deprecate add_node! add_vertex!
 
 function ContinuousSpace(extent, spacing; kwargs...)
     @warn "Specifying `spacing` by position is deprecated. Use keyword `spacing` instead."