+++ src/submodules/pathfinding/astar_continuous.jl
@@ -123,6 +123,7 @@ function Agents.move_along_route!(
         end
         dir = dir ./ dist_to_target
         next_pos = from .+ dir .* (speed * dt)
+        next_pos = Agents.normalize_position(next_pos, model)
         # overshooting means we reached the waypoint
         dist_to_next = euclidean_distance(from, next_pos, model)
         if dist_to_next > dist_to_target